Added the "environment.yml" and instructions for setting up the environment
Sometimes installing packages can be complicated or difficult. I added the "environment.yml" from the environment I made for this class. (This can be used to create a similar environment and works across OS'es!) I also added instructions to the "installing.md" in case anyone has trouble using this or installing packages in the future!